Renamed config target to `zincReportCachedProblems`  and cleanup
